refactor : dao interface 추가로 인한 리팩토링 (기능 가독성 및 인터페이스 규약 설정)
